+++ apps/web/src/components/tx-flow/actions/Execute/__tests__/ExecuteForm.test.tsx
@@ -3,7 +3,7 @@ import { type AsyncResult } from '@safe-global/utils/hooks/useAsync'
 import { createMockSafeTransaction } from '@/tests/transactions'
 import { OperationType } from '@safe-global/safe-core-sdk-types'
 import { type ReactElement } from 'react'
-import { ExecuteForm } from '@/components/tx/SignOrExecuteForm/ExecuteForm'
+import { ExecuteForm } from '../ExecuteForm'
 import * as useGasLimit from '@/hooks/useGasLimit'
 import * as useIsValidExecution from '@/hooks/useIsValidExecution'
 import * as useWalletCanRelay from '@/hooks/useWalletCanRelay'
