+++ apps/web/src/components/tx-flow/actions/Execute/styles.module.css
@@ -1,0 +1,14 @@
+.params {
+  margin-bottom: var(--space-2);
+}
+
+.noBottomBorderRadius :global(.MuiPaper-root) {
+  border-bottom-left-radius: 0 !important;
+  border-bottom-right-radius: 0 !important;
+}
+
+.noTopBorder > div {
+  margin-top: -1px;
+  border-top-left-radius: 0 !important;
+  border-top-right-radius: 0 !important;
+}
