+++ apps/web/src/components/tx/SignOrExecuteForm/SignOrExecuteFormV2.tsx
@@ -5,7 +5,7 @@ import useSafeInfo from '@/hooks/useSafeInfo'
 import { type ReactElement, type ReactNode, useContext, useCallback } from 'react'
 import madProps from '@/utils/mad-props'
 import { useImmediatelyExecutable, useValidateNonce } from './hooks'
-import ExecuteForm from './ExecuteForm'
+import ExecuteForm from '@/components/tx-flow/actions/Execute/ExecuteForm'
 import SignFormV2 from './SignFormV2'
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { useAppSelector } from '@/store'
