+++ apps/web/src/components/tx/SignOrExecuteForm/styles.module.css
@@ -43,17 +43,6 @@
   margin-bottom: var(--space-2);
 }
 
-.noBottomBorderRadius :global(.MuiPaper-root) {
-  border-bottom-left-radius: 0 !important;
-  border-bottom-right-radius: 0 !important;
-}
-
-.noTopBorder > div {
-  margin-top: -1px;
-  border-top-left-radius: 0 !important;
-  border-top-right-radius: 0 !important;
-}
-
 .mobileTxCheckMessages,
 .mobileTxCheckMessages:empty {
   display: none;
